REFACTOR: Set color as value
There's a hex color named 0xFF0094FF which is similar to light blue and it was written like "tint = Color(0xFF0094FF)". So for better color handling, I just set the value of the color in Color.kt file and wrote again like "tint = LightBlue".

From now on, all the colors will be handled like this.