WIP: Set the modalDrawerSheet on top of the stack
Changed the settings to let the modalDrawerSheet to be on top of the
stack. So now, whenever I tap on the menu button from the homescreen,
the modalDrawerSheet will come out on top of it and when I go to other
screens, the topAppBar will be invisible.

And now, need to implement the navigation again cuz it makes some errors
again due to the structure modification.