REFACTOR: Modify transition animation
Found a way to disable the default cross fade transition animation with
navigation dependency. However, this one still needs to be discussed
with the designer so for now just left all the dependency changes and
wrote the additional codes in MainNavGraph.kt with annotation in the
parameter parts of NavHost.